committing PROD values in config/base.ini rather than having something invalid in the repo
I would rather not have machine specific values in files that are under git control but for this one I'm not going to sweat it.  It just means that on STAGE or DEV the values will be modified and developers will need to be careful not to commit base.ini with bad values.  Perhaps adding it to gitignore is warranted